+++ wasm-bridge/bridge.ts
@@ -3,7 +3,7 @@
 
 import type { BridgeBase, InitFn, InitPromise, WasmBaseInstance, WasmImports } from './types.ts';
 
-import { stringToU8a, u8aToString } from 'https://deno.land/x/polkadot@0.2.41/util/mod.ts';
+import { stringToU8a, u8aToString } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { Wbg } from './wbg.ts';
 
@@ -12,7 +12,7 @@ import { Wbg } from './wbg.ts';
  * @description
  * Creates a bridge between the JS and WASM environments.
  *
- * For any bridge it is passed an function white is then called internally at the
+ * For any bridge it is passed an function which is then called internally at the
  * time of initialization. This affectively implements the layer between WASM and
  * the native environment, providing all the plumbing needed for the Wbg classes.
  */