+++ wasm-bridge/types.ts
@@ -1,7 +1,7 @@
 
 export type WasmImports = Record<string, (...args: never[]) => unknown>;
 
-export type WasmMemory = { buffer: ArrayBuffer };
+export interface WasmMemory { buffer: ArrayBuffer }
 
 export declare interface InitResult<C extends WasmBaseInstance> {
   error: string | null;
@@ -29,13 +29,13 @@ export interface BridgeBase<C extends WasmBaseInstance> extends InitResult<C> {
   resultString (): string;
 }
 
-export interface WasmBaseInstance {
-  memory: WasmMemory;
-
-  // wbindgen functions (required and used internally)
-
+export interface WasmBindGen {
   __wbindgen_exn_store (a: number): void;
   __wbindgen_free (a: number, b: number): void;
   __wbindgen_malloc (a: number): number;
   __wbindgen_realloc (a: number, b: number, c: number): number;
+}
+
+export interface WasmBaseInstance extends WasmBindGen {
+  memory: WasmMemory;
 }